not removing files in no-threads test if some outputs differ

+++ tests/no-threads.sh
@@ -9,15 +9,22 @@ echo '(lambda (args) 42)' > $LISP # <- to be dumped as such, so program exits wi
 
 # check that all compile silently and produce equal code
 $@ --run $LISP   # interpret and run
-echo "interpret: $?"
+RINT=$?
+echo "interpret: $RINT"
 
 $@ --no-threads -o $FASL $LISP
 $@ --load $FASL
-echo "fasl: $?"
+RFASL=$?
+echo "fasl: $RFASL"
 
 $@ --no-threads -o $C $LISP
 gcc -o $OBJ $C
 $OBJ
-echo "binary: $?"
+RBIN=$?
+echo "binary: $RBIN"
+
+# don't remove files if there were differences in return values
+test "$RBIN" = "$RFASL" && test "$RFASL" = "$RINT" || exit 1
 
 rm $LISP $FASL $C $OBJ
+
